Rename project and update READMEs towards porting LibTreeSitter to F#
