+++ .circleci/config.yml
@@ -170,8 +170,8 @@ jobs:
           name: Setup backend build environment
           command: |
             set -x
-            if [[ -f /home/dark/tree-sitter.so && ! -f /home/dark/app/backend/src/LibTreeSitter/tree-sitter.so ]]; then
-              mv /home/dark/tree-sitter.so /home/dark/app/backend/src/LibTreeSitter/
+            if [[ -f /home/dark/tree-sitter.so && ! -f /home/dark/app/backend/src/LibTreeSitter.CSharp/tree-sitter.so ]]; then
+              mv /home/dark/tree-sitter.so /home/dark/app/backend/src/LibTreeSitter.CSharp/
               else
               echo "No action needed."
             fi
@@ -179,8 +179,8 @@ jobs:
       - persist_to_workspace:
           root: "."
           paths:
-            - backend/src/LibTreeSitter.Darklang
-            - backend/src/LibTreeSitter
+            - backend/src/LibTreeSitter.CSharp.Darklang
+            - backend/src/LibTreeSitter.CSharp
       - slack-notify-job-failure
 
   # Build server binaries and run tests
