--- backend/src/LibTSParser_FS/README.md
@@ -1,3 +1,0 @@
-WIP of porting LibTreeSitter (a fork of `dotnet-tree-sitter`) from C# to F#.
-
-as of time of writing, this project is unreferenced (so nothing tries to compile it).
