+++ backend/src/LibTreeSitter.CSharp.Darklang/LibTreeSitter.CSharp.Darklang.csproj
@@ -5,13 +5,13 @@
     </PropertyGroup>
 
     <PropertyGroup>
-        <PackageId>LibTreeSitter.Darklang</PackageId>
+        <PackageId>LibTreeSitter.CSharp.Darklang</PackageId>
         <Version>1.0.0</Version>
         <Authors>Darklang</Authors>
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\LibTreeSitter\LibTreeSitter.csproj" />
+      <ProjectReference Include="..\LibTreeSitter.CSharp\LibTreeSitter.CSharp.csproj" />
     </ItemGroup>
 
     <ItemGroup>
