+++ backend/src/LibTreeSitter.CSharp/.gitignore
