--- backend/src/LibTreeSitter.Darklang/README.md
@@ -1,3 +1,0 @@
-This library provides helper functions _specific to Darklang_, wrapping around the
-bindings for [tree-sitter](https://tree-sitter.github.io/tree-sitter/) provided by
-the adjacent `LibTreeSitter` project.
